task2-API/python/requirements.txt
requests==2.32.3
python-dotenv==1.0.1
psycopg[binary]==3.2.3
pymongo==4.8.0